move cli to bin/
