+++ bin/cli.js
