+++ bin/serve.js
@@ -80,17 +80,20 @@ function handleRequest(req, serveDirectory) {
   // get location of hyperparam assets
   const hyperparamPath = decodeURIComponent(import.meta.url)
     .replace('file://', '')
-    .replace('/src/serve.js', '')
+    .replace('/bin/serve.js', '')
 
   if (pathname === '/' || pathname === '/files/') {
     // redirect to /files
     return { status: 301, content: '/files' }
   } else if (pathname.startsWith('/files')) {
     // serve index.html
-    return handleStatic(`${hyperparamPath}/public/index.html`)
-  } else if (pathname.startsWith('/public/')) {
+    return handleStatic(`${hyperparamPath}/dist/index.html`)
+  } else if (pathname.startsWith('/assets/')) {
     // serve static files
-    return handleStatic(`${hyperparamPath}${pathname}`)
+    return handleStatic(`${hyperparamPath}/dist${pathname}`)
+  // else if (pathname.startsWith('/public/')) {
+  //   // serve static files
+  //   return handleStatic(`${hyperparamPath}${pathname.replace(/^(\/public).*/, '/dist')}`)
   } else if (serveDirectory && pathname === '/api/store/list') {
     // serve file list
     const prefix = parsedUrl.query.prefix || ''
