+++ bin/streamConverters.js
