+++ eslint.config.js
@@ -100,4 +100,21 @@ export default tseslint.config(
     files: ['**/*.js'],
     ...tseslint.configs.disableTypeChecked,
   },
+  // {
+  //   files: ['bin/**/*.js'],
+  //   plugins: {
+  //     typescript,
+  //   },
+  //   languageOptions: {
+  //     globals: {
+  //       ...globals.browser,
+  //       ...globals.node,
+  //     },
+  //   },
+  //   rules: {
+  //     ...javascript.configs.recommended.rules,
+  //     ...typescript.configs.recommended.rules,
+  //     ...sharedJsRules,
+  //   },
+  // }
 )
