+++ index.html
@@ -5,12 +5,12 @@
     <title>hyperparam</title>
     <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Mulish:wght@400;600&display=swap"/>
     <link href="favicon.png" rel="icon" />
-    <meta name="description" content="demo for the Hyperparam components" />
+    <meta name="description" content="hyperparam is the missing UI for machine learning" />
     <meta name="theme-color" content="#6b00ff">
     <meta name="viewport" content="width=device-width, initial-scale=1" />
   </head>
   <body>
-    <div id="root"></div>
-    <script type="module" src="./src/demo.tsx"></script>
+    <div id="app"></div>
+    <script type="module" src="./src/app.tsx"></script>
   </body>
 </html>
