+++ package.json
@@ -1,18 +1,13 @@
 {
-  "name": "@hyparam/components",
-  "version": "0.1.19",
-  "description": "React components for hyparam apps",
-  "keywords": [
-    "component",
-    "data",
-    "javascript",
-    "react"
-  ],
+  "name": "hyperparam",
+  "version": "0.2.23",
+  "description": "Hyperparam CLI",
+  "author": "Hyperparam",
+  "homepage": "https://hyperparam.app",
   "license": "MIT",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/hyparam/hyperparam-cli.git",
-    "directory": "packages/components"
   },
   "type": "module",
   "types": "lib/index.d.ts",
@@ -29,13 +24,17 @@
   ],
   "scripts": {
     "build": "tsc -b && vite build",
-    "coverage": "vitest run --coverage --coverage.include=src",
-    "dev": "vite",
+    "coverage": "vitest run --coverage --coverage.include=src --coverage.include=bin",
+    "dev:vite": "vite",
+    "dev:serve": "npm run serve",
+    "dev:url": "npm run dev:serve -- https://hyperparam.blob.core.windows.net/hyperparam/starcoderdata-js-00000-of-00065.parquet",
     "lint": "eslint .",
     "prepublishOnly": "npm run build",
+    "preview:vite": "vite preview",
+    "serve": "node bin/cli.js",
+    "preserve": "npm run build",
     "test": "vitest run",
-    "typecheck": "tsc --noEmit",
-    "preview": "vite preview"
+    "typecheck": "tsc --noEmit"
   },
   "dependencies": {
     "hightable": "0.11.0",
