--- packages/cli/public/index.html
@@ -1,17 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8">
-    <title>hyperparam</title>
-    <link rel="stylesheet" href="./public/styles.css">
-    <link rel="stylesheet" href="https://fonts.googleapis.com/css2?family=Mulish:wght@400;600&display=swap"/>
-    <link href="./public/favicon.png" rel="icon" />
-    <meta name="description" content="hyperparam is the missing UI for machine learning" />
-    <meta name="theme-color" content="#6b00ff">
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-  </head>
-  <body>
-    <div id="app"></div>
-    <script type="module" src="./public/build/app.min.js"></script>
-  </body>
-</html>
