+++ packages/cli/tsconfig.json
@@ -9,5 +9,5 @@
     "resolveJsonModule": true,
     "strict": true
   },
-  "include": ["src", "test"]
+  "include": ["bin", "test"]
 }
