+++ src/styles/app.css
@@ -1,35 +1,49 @@
 * {
   box-sizing: border-box;
+  font-family: 'Mulish', 'Helvetica Neue', Helvetica, Arial, sans-serif;
   margin: 0;
   padding: 0;
 }
 
 #app {
   display: flex;
-  font-family: 'Mulish', 'Helvetica Neue', Helvetica, Arial, sans-serif;
   flex-direction: column;
   height: 100vh;
 }
 
+/* table overrides */
+.table-container .table-corner {
+  background: url('https://hyperparam.app/assets/table/hyperparam.svg') #f9f4ff no-repeat center 6px;
+}
+main .table-container .table th {
+  background-color: #f1f1f3;
+}
+main .table td {
+  cursor: pointer;
+}
+main .table tbody tr:hover th:first-child {
+  background-color: #ccd;
+  border-right: 1px solid #ccc;
+}
+
 /* brand logo */
 .brand {
-  color: #111;
-  display: flex;
   align-items: center;
-  filter: drop-shadow(0 0 2px #444);
+  color: #222;
+  display: flex;
+  filter: drop-shadow(0 0 2px #bbb);
   font-family: 'Century Gothic', 'Helvetica Neue', Helvetica, Arial, sans-serif;
   font-size: 1.1em;
   font-weight: bold;
   text-orientation: mixed;
-  opacity: 0.85;
+  letter-spacing: 0.3px;
   padding: 10px 12px;
   user-select: none;
   writing-mode: vertical-rl;
 }
 .brand:hover {
-  color: #fff;
-  filter: drop-shadow(0 0 2px #333);
-  opacity: 0.9;
+  color: #222;
+  filter: drop-shadow(0 0 2px #afa6b9);
   text-decoration: none;
 }
 .brand::before {
@@ -50,6 +64,9 @@ a:hover {
   color: #000;
   text-decoration: underline;
 }
+button {
+  cursor: pointer;
+}
 
 /* layout */
 main {
@@ -58,15 +75,44 @@ main {
   max-width: 100vw;
 }
 
+.spinner {
+  border: 6px solid #e8e8e8;
+  border-top: 6px solid #68a;
+  border-radius: 50%;
+  width: 32px;
+  height: 32px;
+  animation: spin 1s linear infinite, fadeIn 3s, ease-in-out forwards;
+}
+@keyframes spin {
+  0% { transform: rotate(0deg); }
+  100% { transform: rotate(360deg); }
+}
+@keyframes fadeIn {
+  0% { opacity: 0; }
+  100% { opacity: 1; }
+}
+.center {
+  flex: 1;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
 /* sidebar */
 .nav {
   display: flex;
   flex-direction: column;
   height: 100vh;
   width: 48px;
-  background-image: linear-gradient(to bottom, #667, #585669);
-  box-shadow: 0 0 4px rgba(10, 10, 10, 0.5);
   height: 100vh;
+}
+.nav > div {
+  background-image: linear-gradient(to bottom, #f2f2f2, #e4e4e4);
+  box-shadow: 0 0 6px rgba(10, 10, 10, 0.4);
+  height: 100vh;
+  position: absolute;
+  width: 48px;
+  z-index: 30;
 }
 
 /* content area */
@@ -90,20 +136,49 @@ main {
 /* top navbar */
 .top-header {
   align-items: center;
-  background: linear-gradient(to right, #353540, #24202b);
-  color: #dde4ea;
   display: flex;
+  font-size: 18px;
   height: 32px;
   justify-content: space-between;
   min-height: 32px;
-  padding-left: 8px;
+  padding-left: 20px;
+  padding-right: 10px;
 }
 .top-header h1 {
   font-size: 18px;
-  margin: 0;
+  margin: 4px 0 0 0; /* top */
+  user-select: none;
 }
-.top-header a {
-  color: #e0e8ee;
+
+.top-header-divided {
+  border-bottom: 1px solid #ddd;
+  background: #eee;
+}
+
+.top-actions {
+  left: auto;
+}
+/* search */
+input.search {
+  background: #fff url("../assets/search.svg") no-repeat center right 8px;
+  border: 1px solid transparent;
+  border-radius: 8px;
+  flex-shrink: 1;
+  font-size: 12px;
+  height: 24px;
+  min-width: 0;
+  outline: none;
+  padding: 4px 20px 2px 8px;
+  width: 100px;
+  transition-duration: 0.3s;
+  transition-property: border, width;
+}
+input.search:focus {
+  border-color: #778;
+  box-shadow: 0 0 1px #556;
+  color: #444;
+  padding-left: 8px;
+  width: 180px;
 }
 
 /* file path */
@@ -117,8 +192,9 @@ main {
   display: none;
 }
 .path a {
-  color: #f0f8ff;
+  color: #222622;
   font-family: 'Courier New', Courier, monospace;
+  font-weight: 600;
   font-size: 18px;
   text-overflow: ellipsis;
   white-space: nowrap;
@@ -169,10 +245,12 @@ main {
   max-height: 0;
   padding: 0;
   background-color: #dd111199;
-  font-family: monospace;
   overflow: hidden;
   transition: max-height 0.3s;
   white-space: pre-wrap;
+}
+.error-bar * {
+  font-family: monospace;
 }
 .show-error {
   max-height: 30%;
@@ -183,7 +261,7 @@ main {
   justify-content: space-between;
   min-height: 100%;
   overflow-y: auto;
-  padding: 10px;
+  padding: 10px 10px 10px 20px;
 }
 
 /* error bar close button */
@@ -220,6 +298,7 @@ button.close-button:hover {
   flex: 1;
   list-style: none;
   overflow-y: auto;
+  padding-left: 0;
   /* browsers like to cover the bottom row */
   padding-bottom: 24px;
 }
@@ -230,10 +309,10 @@ button.close-button:hover {
   border-top: none;
 }
 .file-list a {
-  border-top: 1px solid #bbb;
+  border-top: 1px solid #ddd;
   color: #444;
   display: flex;
-  padding: 8px 12px;
+  padding: 8px 16px 8px 20px;
   text-decoration: none;
 }
 .file-list a:hover {
@@ -282,12 +361,13 @@ button.close-button:hover {
 }
 .view-header {
   align-items: center;
-  background-color: #ccc;
+  background-color: #f2f2f2;
   color: #444;
   display: flex;
   gap: 16px;
   height: 24px;
-  padding: 0 8px;
+  overflow: hidden;
+  padding: 0 16px;
   /* all one line */
   text-overflow: ellipsis;
   white-space: nowrap;
@@ -298,6 +378,7 @@ button.close-button:hover {
   color: #d6d6d6;
   display: block;
   flex: 1;
+  font-family: monospace;
   padding: 8px;
   width: 100%;
   overflow-x: auto;
@@ -360,6 +441,7 @@ button.close-button:hover {
   display: flex;
   flex: 1;
   flex-direction: column;
+  min-width: 100px;
   overflow: auto;
 }
 
@@ -368,6 +450,10 @@ button.close-button:hover {
   flex-direction: column;
   width: 0;
   transition: width 0.2s;
+}
+
+.slideDragging {
+  transition: none;
 }
 
 .resizer {
@@ -389,17 +475,18 @@ button.close-button:hover {
 .slideClose:hover {
   background: none;
   border: none;
-  color: inherit;
-  margin-left: auto;
+  color: #888;
+  font-size: 16px;
   height: 24px;
+  margin-right: auto;
   outline: none;
-  padding: 0;
+  transition: color 0.3s;
 }
 .slideClose::before {
   content: "\27E9\27E9";
 }
 .slideClose:hover {
-  font-weight: bold;
+  color: #000;
 }
 
 /* viewers */
